# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PlayZone Bar API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reservations:
    get:
      operationId: listReservations
      tags: [reservations]
      summary: List reservations
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, cancelled, completed, checkin]
        - name: type
          in: query
          schema:
            type: string
            enum: [salao, gamer]
        - name: date
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of reservations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservationList"
    post:
      operationId: createReservation
      tags: [reservations]
      summary: Create a reservation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReservationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
  /reservations/{id}:
    get:
      operationId: getReservation
      tags: [reservations]
      summary: Get a reservation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reservation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
        "404":
          description: Not found
    patch:
      operationId: updateReservation
      tags: [reservations]
      summary: Update a reservation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReservationUpdate"
      responses:
        "200":
          description: Updated reservation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
    delete:
      operationId: deleteReservation
      tags: [reservations]
      summary: Delete a reservation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /reservations/{id}/checkin:
    post:
      operationId: checkInReservation
      tags: [reservations]
      summary: Check in a reservation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Checked in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
  /reservations/availability:
    get:
      operationId: getAvailability
      tags: [reservations]
      summary: Get available time slots for a date and type
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
        - name: type
          in: query
          required: true
          schema:
            type: string
            enum: [salao, gamer]
      responses:
        "200":
          description: Available slots
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilitySlots"

  # WAITLIST
  /waitlist:
    get:
      operationId: listWaitlist
      tags: [waitlist]
      summary: List waitlist entries
      responses:
        "200":
          description: Waitlist entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WaitlistEntry"
    post:
      operationId: createWaitlistEntry
      tags: [waitlist]
      summary: Add to waitlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WaitlistInput"
      responses:
        "201":
          description: Added to waitlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WaitlistEntry"
  /waitlist/{id}:
    delete:
      operationId: deleteWaitlistEntry
      tags: [waitlist]
      summary: Remove from waitlist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed

  # GALLERY
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List admin users
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
    post:
      operationId: createUser
      tags: [users]
      summary: Create admin user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
  /users/{id}:
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    Reservation:
      type: object
      required: [id, name, phone, people, date, time, type, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        people:
          type: integer
        date:
          type: string
        time:
          type: string
        type:
          type: string
          enum: [salao, gamer]
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed, checkin]
        notes:
          type: ["string", "null"]
        couponCode:
          type: ["string", "null"]
        discountApplied:
          type: ["number", "null"]
        qrCode:
          type: ["string", "null"]
        checkedInAt:
          type: ["string", "null"]
        createdAt:
          type: string
    ReservationList:
      type: object
      required: [reservations, total, page, limit]
      properties:
        reservations:
          type: array
          items:
            $ref: "#/components/schemas/Reservation"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    ReservationInput:
      type: object
      required: [name, phone, people, date, time, type]
      properties:
        name:
          type: string
        phone:
          type: string
        people:
          type: integer
          minimum: 1
        date:
          type: string
        time:
          type: string
        type:
          type: string
          enum: [salao, gamer]
        notes:
          type: string
        couponCode:
          type: string
    ReservationUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        people:
          type: integer
        date:
          type: string
        time:
          type: string
        type:
          type: string
          enum: [salao, gamer]
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed, checkin]
        notes:
          type: string
    AvailabilitySlots:
      type: object
      required: [date, type, slots]
      properties:
        date:
          type: string
        type:
          type: string
        slots:
          type: array
          items:
            type: string
    WaitlistEntry:
      type: object
      required: [id, name, phone, date, time, type, people, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        date:
          type: string
        time:
          type: string
        type:
          type: string
          enum: [salao, gamer]
        people:
          type: integer
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    WaitlistInput:
      type: object
      required: [name, phone, date, time, type, people]
      properties:
        name:
          type: string
        phone:
          type: string
        date:
          type: string
        time:
          type: string
        type:
          type: string
          enum: [salao, gamer]
        people:
          type: integer
        notes:
          type: string
...
    AdminUser:
      type: object
      required: [id, email, name, role, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
        createdAt:
          type: string
    UserInput:
      type: object
      required: [email, name, password, role]
      properties:
        email:
          type: string
        name:
          type: string
        password:
          type: string
        role:
          type: string
    UserUpdate:
      type: object
      properties:
        email:
          type: string
        name:
          type: string
        password:
          type: string
        role:
          type: string